Reduce the changes in our custom metro config (#4939)
* Reduce need for custom metro configs

* Change files

* Fix for integration tests
