+++ change/@office-iss-react-native-win32-2020-05-18-09-46-04-metrofix.json
@@ -1,0 +1,8 @@
+{
+  "type": "prerelease",
+  "comment": "Reduce need for custom metro configs",
+  "packageName": "@office-iss/react-native-win32",
+  "email": "acoates@microsoft.com",
+  "dependentChangeType": "patch",
+  "date": "2020-05-18T16:46:03.284Z"
+}
