+++ change/react-native-windows-2020-05-18-09-46-04-metrofix.json
@@ -1,0 +1,8 @@
+{
+  "type": "prerelease",
+  "comment": "Reduce need for custom metro configs",
+  "packageName": "react-native-windows",
+  "email": "acoates@microsoft.com",
+  "dependentChangeType": "patch",
+  "date": "2020-05-18T16:46:04.876Z"
+}
