+++ packages/react-native-win32/react-native.config.js
@@ -6,6 +6,7 @@ module.exports = {
         linkConfig: () => null,
         projectConfig: (projectRoot, projectParams) => null,
         dependencyConfig: (projectRoot, dependencyParams) => null,
+        npmPackageName: '@office-iss/react-native-win32',
       },
     },
   };
