+++ vnext/metro.config.js
@@ -3,13 +3,34 @@
  */
 const fs = require('fs');
 const path = require('path');
-
-const {
-  getModulesRunBeforeMainModule,
-  reactNativePlatformResolver,
-} = require('react-native-windows/metro-react-native-platform');
-
 const rnwPath = __dirname;
+const {resolve} = require('metro-resolver');
+
+function reactNativePlatformResolver(platformImplementations) {
+  return (context, _realModuleName, platform, moduleName) => {
+    let backupResolveRequest = context.resolveRequest;
+    delete context.resolveRequest;
+
+    try {
+      let modifiedModuleName = moduleName;
+      if (platformImplementations[platform]) {
+        if (moduleName === 'react-native') {
+          modifiedModuleName = platformImplementations[platform];
+        } else if (moduleName.startsWith('react-native/')) {
+          modifiedModuleName = `${
+            platformImplementations[platform]
+          }/${modifiedModuleName.slice('react-native/'.length)}`;
+        }
+      }
+      let result = resolve(context, modifiedModuleName, platform);
+      return result;
+    } catch (e) {
+      throw e;
+    } finally {
+      context.resolveRequest = backupResolveRequest;
+    }
+  };
+}
 
 module.exports = {
   // WatchFolders is only needed due to the yarn workspace layout of node_modules, we need to watch the symlinked locations separately
@@ -19,6 +40,7 @@ module.exports = {
   ],
 
   resolver: {
+    // We need a custom resolveRequest right now since our integration tests use a "windesktop" platform thats specific to integration tests.
     resolveRequest: reactNativePlatformResolver({
       windesktop: 'react-native-windows',
       windows: 'react-native-windows',
@@ -29,14 +51,7 @@ module.exports = {
     },
   },
 
-  serializer: {
-    getModulesRunBeforeMainModule,
-  },
-
   transformer: {
-    // The cli defaults this to a full path to react-native, which bypasses the reactNativePlatformResolver above
-    // Hopefully we can fix the default in the future
-    assetRegistryPath: 'react-native/Libraries/Image/AssetRegistry',
     getTransformOptions: async () => ({
       transform: {
         experimentalImportSupport: false,
